+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Features
 - Add X-Up-Context response header and put_context/2 helper for updating layer context.
 - Add root?/1 and overlay?/1 helpers for layer detection.
 - Add X-Up-Open-Layer response header and open_layer/2 helper for forcing overlay layers.
+- Add origin_mode/1 and fail_context/1 helpers for experimental request headers.
 
 ## v1.3.0 (2021-06-17)